use template method pattern for printing dts nodes
AbstractNode is an abstract base class that implements the base functionality for a node, but can't be instantiated.
It's __str__ method handles printing the name/label, and the braces; subclasses must implement the _print_properties method, which
prints the property strings that get inserted into the device tree node string.

Files were moved into a package, and naming conventions were updated to reflect PEP 8 conventions
