+++ device_tree_overlays/dtogen/__init__.py
@@ -1,0 +1,5 @@
+__title__ = 'dtogen'
+__author__ = 'Trevor Vannoy <vannoy@flatearthinc.com>'
+__license__ = 'GPL v3'
+__copyright__ = 'Copyright 2020 Flat Earth Inc.'
+__all__ = ('DeviceTreeNode', 'FpgaRegionNode')
